AWS/AWS/aws_container_todo/app/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
mysql-connector-python>=8.2.0
pydantic>=2.0.0
